backend/requirements.txt
fastapi
uvicorn
python-dotenv
opentelemetry-distro
opentelemetry-exporter-otlp
aiohttp
httpx
modelcontextprotocol
elasticsearch
